+++ R/mission_orbits.R
@@ -1,5 +1,202 @@
 
-utils::globalVariables(c('i'))                        # for the foreach for-loop
+utils::globalVariables(c('i',                        # for the foreach for-loop
+                         'Type'))                    # for the Type column when I use the base R subset() function
+
+
+#' Extracting the download url's from the Technical Specs website
+#'
+#' @param technical_specs_url a character string specifying the technical specs website
+#' @param verbose a boolean. If TRUE then information will be printed out in the console
+#'
+#' @return a 'data.table' object
+#'
+#' @references
+#'
+#' https://icesat-2.gsfc.nasa.gov/science/specs
+#' 
+#' @importFrom rvest read_html html_elements       
+#' @importFrom data.table data.table
+#' @import magrittr
+#' 
+#' @export
+#'
+#' @examples
+#'
+#' \dontrun{
+#'
+#' require(IceSat2R)
+#' 
+#' orbs = latest_orbits(verbose = TRUE)
+#' orbs
+#' 
+#' }
+
+latest_orbits = function(technical_specs_url = "https://icesat-2.gsfc.nasa.gov/science/specs",
+                         verbose = FALSE) {
+  if (verbose) {
+    t_start = proc.time()
+    cat("Access the data of the technical specs website ...\n")
+  }
+  
+  pg = rvest::read_html(technical_specs_url)
+  
+  if (verbose) cat("Extract the .zip files and the corresponding titles ...\n")
+  url_zipfiles = pg %>%
+    rvest::html_elements("a[href$='.zip']") %>%       # see: https://stackoverflow.com/a/48146333
+    as.character()
+  # rvest::html_text2()                               # returns only the url-titles
+  
+  if (verbose) cat("Keep the relevant data from the url's and titles ...\n")
+  df_out = lapply(seq_along(url_zipfiles), function(x) {
+    item = url_zipfiles[x]
+    item = gsub(pattern = '<a href=\"', replacement = '', x = item)
+    item = gsub(pattern = '</a>', replacement = '', x = item)
+    
+    idx_str = which(gregexpr(pattern = '<strong>', text = item) != -1)
+    if (length(idx_str) > 0) {
+      item = gsub(pattern = '<strong>', replacement = '', x = item)
+      item = gsub(pattern = '</strong>', replacement = '', x = item)
+    }
+    
+    item = strsplit(x = item, split = '[\">]')[[1]]
+    item = item[item != ""]
+    item = trimws(x = item, which = 'both')
+    item
+  })
+  
+  df_out = do.call(rbind, df_out)
+  
+  if (verbose) cat("Process the nominal and time specific orbits separately ...\n")
+  idx_time_spec = which(gregexpr(pattern = "^RGTs", text = df_out[, 2]) != -1)
+  nominal_orbits = df_out[-idx_time_spec, ]
+  time_specific = df_out[idx_time_spec, ]
+  
+  if (verbose) cat("Adjust the Dates of the time specific orbits ...\n")
+  dates_spec = as.vector(unlist(lapply(strsplit(x = time_specific[, 2], split = "for"), function(x) trimws(x[2], which = 'both'))))
+  dates_spec = strsplit(x = dates_spec, split = '-')
+  dates_spec = lapply(dates_spec, function(x) {
+    x = trimws(x = x, which = 'both')
+    x = gsub(pattern = "[(]", replacement = ",", x = x)
+    x = gsub(pattern = "[)]", replacement = "", x = x)
+    x = strsplit(x = x, split = ',')
+    
+    x_out = rep(x = NA_character_, 2)
+    
+    for (i in seq_along(x)) {
+      y = trimws(x = x[[i]], which = 'both')
+      if (length(y) == 1) {
+        y = c(y, trimws(x = x[[i+1]][2], which = 'both'))
+      }
+      x_out[i] = paste(y, collapse = ' ')
+    }
+    x_out
+  })
+  
+  dates_spec = lapply(dates_spec, function(x) {
+    as.character(as.Date(x, format = "%b %d %Y"))
+  })
+  
+  dates_spec = do.call(rbind, dates_spec)
+  
+  if (verbose) cat("Create the nominal orbits data.table ...\n")
+  nominal_orbits = data.table::data.table(nominal_orbits, stringsAsFactors = F)
+  colnames(nominal_orbits) = c('Url', 'RGT')
+  nominal_orbits$Date_from = NA_character_
+  nominal_orbits$Date_to = NA_character_
+  nominal_orbits$Type = 'Nominal'
+  
+  if (verbose) cat("Create the time specific orbits data.table ...\n")
+  time_specific = data.table::data.table(time_specific[, 1, drop = F], stringsAsFactors = F)
+  colnames(time_specific) = 'Url'
+  time_specific$RGT = trimws(gsub(pattern = '.zip', '', basename(path = time_specific$Url)), which = 'both')
+  time_specific$Date_from = dates_spec[, 1]
+  time_specific$Date_to = dates_spec[, 2]
+  time_specific$Type = 'Time_Specific'
+  
+  if (verbose) cat("Return a single data.table ...\n")
+  res_both = rbind(nominal_orbits, time_specific)
+  if (verbose) compute_elapsed_time(time_start = t_start)
+  
+  return(res_both)
+}
+
+
+
+#' Nominal mission orbits
+#'
+#' @param orbit_area either NULL or a character string specifying the earth partition to use, it can be one of 'antarctic', 'arctic', 'western_hemisphere' and 'eastern_hemisphere'
+#' @param technical_specs_url a character string specifying the technical specs website
+#' @param verbose a boolean. If TRUE then information will be printed out in the console
+#'
+#' @return a list object with the available nominal orbits
+#'
+#' @importFrom glue glue
+#' @importFrom data.table fread
+#'
+#' @references
+#'
+#' https://icesat-2.gsfc.nasa.gov/science/specs
+#'
+#' @export
+#'
+#' @examples
+#'
+#' require(IceSat2R)
+#'
+#' #.............................
+#' # all available nominal orbits
+#' #.............................
+#'
+#' nomin_orb = available_nominal_orbits(verbose = TRUE)
+#' nomin_orb
+#'
+#' #........................
+#' # specific nominal orbits
+#' #........................
+#'
+#' nomin_orb = available_nominal_orbits(orbit_area = 'arctic',
+#'                                      verbose = TRUE)
+#' nomin_orb
+#'
+
+available_nominal_orbits = function(orbit_area = NULL,
+                                    technical_specs_url = "https://icesat-2.gsfc.nasa.gov/science/specs",
+                                    verbose = FALSE) {
+
+  if (verbose) cat(glue::glue("The available Icesat-2 orbits will be red from '{technical_specs_url}' ..."), '\n')
+  avail_urls = tryCatch(latest_orbits(technical_specs_url = technical_specs_url, verbose = verbose), error = function(e) e)
+  if (inherits(avail_urls, "error")) {
+    if (verbose) message(glue::glue("The 'latest_orbits()' function gave the following error: '{avail_urls$message}'! The existing orbits file will be uploaded!"))
+    avail_urls = data.table::fread(file = system.file('data_files', 'technical_specs_urls.csv', package = "IceSat2R"), stringsAsFactors = F, header = T)
+  }
+  
+  nom_orb = subset(avail_urls, Type == 'Nominal')
+  
+  rgts = as.vector(unlist(lapply(strsplit(x = nom_orb$RGT, split = ' '), function(x) {
+    if (length(x) == 2) {
+      x = x[1]
+    }
+    else {
+      x = x[1:2]
+    }
+    x = paste(x, collapse = '_')
+    tolower(x)
+  })))
+  
+  nom_orb$RGT = rgts
+  intern_orbits = list()
+  for (i in 1:nrow(nom_orb)) {
+    intern_orbits[[nom_orb$RGT[i]]] = nom_orb$Url[i]
+  }
+
+  if (!is.null(orbit_area)) {
+    if (!all(orbit_area %in% names(intern_orbits))) stop(glue::glue("The 'orbit_area' parameter can be one of {paste(rgts, collapse = ', ')}!"), call. = F)
+    intern_orbits = intern_orbits[orbit_area]
+  }
+
+  return(intern_orbits)
+}
+
 
 
 #' Overall Mission Orbits
@@ -53,15 +250,12 @@ overall_mission_orbits = function(orbit_area,
     t_start = t_downl = proc.time()
   }
 
-  orbits = list(antarctic = 'https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/antarcticaallorbits.zip',                          # 'antarctic_orbit'
-                arctic = 'https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/arcticallorbits.zip',                                 # 'arctic_orbit'
-                western_hemisphere = 'https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/ICESat2groundtracksWesternHem.zip',       # 'western_hemisphere_ground_tracks'
-                eastern_hemisphere = 'https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/ICESat2groundtracksEasternHem.zip')       # 'eastern_hemisphere_ground_tracks'
+  orbits = available_nominal_orbits(orbit_area = orbit_area, verbose = verbose)
 
   tmp_orbit_file = tempfile(fileext = '.zip')
-  if (verbose) cat(glue::glue("The orbits of '{orbit_area}' will be downloaded in the '{tmp_orbit_file}' file ..."), '\n')
+  if (verbose) cat(glue::glue("The orbits of '{orbit_area}' will be downloaded (temporarily) in the '{tmp_orbit_file}' file ..."), '\n')
   utils::download.file(url = orbits[[orbit_area]], destfile = tmp_orbit_file, method = download_method, quiet = !verbose)
-  tmp_orbits_dir = file.path(tempdir(), glue::glue('orbits_{orbit_area}'))                                                  # create a second directory inside the temporary directory to extract the .zip file
+  tmp_orbits_dir = file.path(tempdir(), glue::glue('orbits_{orbit_area}'))                                # create a second directory inside the temporary directory to extract the .zip file
   if (!dir.exists(tmp_orbits_dir)) dir.create(tmp_orbits_dir)
 
   if (verbose) cat(glue::glue("The downloaded .zip file will be extracted in the '{tmp_orbits_dir}' directory ..."), '\n')
@@ -163,8 +357,13 @@ overall_mission_orbits = function(orbit_area,
 #' Reference Ground Tracks (RGTs)
 #'
 #' @param only_cycle_names a boolean. If TRUE then only the RGT (Reference Ground Track) cycle names will be returned. Otherwise all orbit files, dates and cycle names.
+#' @param technical_specs_url a character string specifying the technical specs website
+#' @param verbose a boolean. If TRUE then information will be printed out in the console
 #'
 #' @return a list object with the available orbit files, dates and cycle names
+#' 
+#' @importFrom glue glue
+#' @importFrom data.table fread
 #'
 #' @references
 #'
@@ -174,63 +373,57 @@ overall_mission_orbits = function(orbit_area,
 #'
 #' @examples
 #'
+#' \dontrun{
+#'
 #' require(IceSat2R)
 #'
 #' #.................................................
 #' # all available orbit files, dates and cycle names
 #' #.................................................
 #'
-#' avail_dat = available_RGTs(only_cycle_names = FALSE)
+#' avail_dat = available_RGTs(only_cycle_names = FALSE, 
+#'                            verbose = TRUE)
 #' avail_dat
 #'
 #' #.............................
 #' # receive only the cycle names
 #' #.............................
 #'
-#' avail_cycles = available_RGTs(only_cycle_names = TRUE)
+#' avail_cycles = available_RGTs(only_cycle_names = TRUE,
+#'                               verbose = TRUE)
 #' avail_cycles
+#' 
+#' }
 
 
-available_RGTs = function(only_cycle_names = FALSE) {
-
-  orbit_files = list(RGT_cycle_1 = 'https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle1_date_time.zip',
-                     RGT_cycle_2 = 'https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle2_date_time.zip',
-                     RGT_cycle_3 = 'https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle3_date_time_0.zip',
-                     RGT_cycle_4 = 'https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle4_date_time_rev2_0.zip',
-                     RGT_cycle_5 = 'https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle5_date_time_rev3.zip',
-                     RGT_cycle_6 = 'https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle6_date_time.zip',
-                     RGT_cycle_7 = 'https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle7_date_time.zip',
-                     RGT_cycle_8 = 'https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle8_date_time.zip',
-                     RGT_cycle_9 = 'https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle9_date_time.zip',
-                     RGT_cycle_10 = 'https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle10_date_time.zip',
-                     RGT_cycle_11 = 'https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle11_date_time.zip',
-                     RGT_cycle_12 = 'https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2RGTscycle12datetime.zip',
-                     RGT_cycle_13 = 'https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle13_date_time.zip',
-                     RGT_cycle_14 = 'https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle14_date_time.zip',
-                     RGT_cycle_15 = 'https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle15_date_time.zip')
-
-  orbit_dates = list(RGT_cycle_1 = seq(from = as.Date('2018-10-13'), to = as.Date('2018-12-28'), by = '1 day'),
-                     RGT_cycle_2 = seq(from = as.Date('2018-12-28'), to = as.Date('2019-03-29'), by = '1 day'),
-                     RGT_cycle_3 = seq(from = as.Date('2019-03-29'), to = as.Date('2019-06-28'), by = '1 day'),
-                     RGT_cycle_4 = seq(from = as.Date('2019-07-09'), to = as.Date('2019-09-26'), by = '1 day'),
-                     RGT_cycle_5 = seq(from = as.Date('2019-09-26'), to = as.Date('2019-12-26'), by = '1 day'),
-                     RGT_cycle_6 = seq(from = as.Date('2019-12-26'), to = as.Date('2020-03-26'), by = '1 day'),
-                     RGT_cycle_7 = seq(from = as.Date('2020-03-26'), to = as.Date('2020-06-25'), by = '1 day'),
-                     RGT_cycle_8 = seq(from = as.Date('2020-06-25'), to = as.Date('2020-09-24'), by = '1 day'),
-                     RGT_cycle_9 = seq(from = as.Date('2020-09-24'), to = as.Date('2020-12-23'), by = '1 day'),
-                     RGT_cycle_10 = seq(from = as.Date('2020-12-24'), to = as.Date('2021-03-24'), by = '1 day'),
-                     RGT_cycle_11 = seq(from = as.Date('2021-03-24'), to = as.Date('2021-06-23'), by = '1 day'),
-                     RGT_cycle_12 = seq(from = as.Date('2021-06-23'), to = as.Date('2021-09-22'), by = '1 day'),
-                     RGT_cycle_13 = seq(from = as.Date('2021-09-22'), to = as.Date('2021-12-22'), by = '1 day'),
-                     RGT_cycle_14 = seq(from = as.Date('2021-12-22'), to = as.Date('2022-03-23'), by = '1 day'),
-                     RGT_cycle_15 = seq(from = as.Date('2022-03-23'), to = as.Date('2022-06-21'), by = '1 day'))
+available_RGTs = function(only_cycle_names = FALSE,
+                          technical_specs_url = "https://icesat-2.gsfc.nasa.gov/science/specs",
+                          verbose = FALSE) {
+  
+  if (verbose) cat(glue::glue("The available Icesat-2 orbits will be red from '{technical_specs_url}' ..."), '\n')
+  avail_urls = tryCatch(latest_orbits(technical_specs_url = technical_specs_url, verbose = verbose), error = function(e) e)
+  if (inherits(avail_urls, "error")) {
+    if (verbose) message(glue::glue("The 'latest_orbits()' function gave the following error: '{avail_urls$message}'! The existing orbits file will be uploaded!"))
+    avail_urls = data.table::fread(file = system.file('data_files', 'technical_specs_urls.csv', package = "IceSat2R"), stringsAsFactors = F, header = T)
+  }
+  
+  time_spec = subset(avail_urls, Type == 'Time_Specific')
+  time_spec$Date_from = as.Date(time_spec$Date_from)
+  time_spec$Date_to = as.Date(time_spec$Date_to)
+  time_spec = time_spec[order(time_spec$Date_from, decreasing = F), ]              # sort by date to have the RGT-cycles ordered in increasing order
+  
+  orbit_files = orbit_dates = list()
+  
+  for (i in 1:nrow(time_spec)) {
+    nam_i = glue::glue("RGT_cycle_{i}")
+    orbit_files[[nam_i]] = time_spec$Url[i]
+    orbit_dates[[nam_i]] = seq(from = time_spec$Date_from[i], to = time_spec$Date_to[i], by = '1 day')
+  }
 
   orb_nams = names(orbit_files)
-  orb_dats = names(orbit_dates)
-  if (any(orb_nams != orb_dats)) stop("There is a mismatch between the names of the 'orbit_files' and 'orbit_dates'!", call. = F)
 
   if (only_cycle_names) {
-    return(names(orbit_files))
+    return(orb_nams)
   }
   else {
     return(list(orbit_files = orbit_files,
@@ -278,7 +471,8 @@ available_RGTs = function(only_cycle_names = FALSE) {
 #' # observe and choose one of the available RGT-cycles
 #' #...................................................
 #'
-#' avail_cycles = available_RGTs(only_cycle_names = TRUE)
+#' avail_cycles = available_RGTs(only_cycle_names = TRUE,
+#'                               verbose = TRUE)
 #' avail_cycles
 #'
 #' #.....................................
@@ -376,7 +570,6 @@ revisit_time_RGTs = function(RGT_cycle = NULL,
 #'                                    download_method = 'curl',
 #'                                    threads = 1,
 #'                                    verbose = TRUE)
-#'
 #' str(res_rgt_one)
 #'
 #' #..........................................................
@@ -411,7 +604,8 @@ revisit_time_RGTs = function(RGT_cycle = NULL,
 #' # Observe the available RGT-cycles and use all available threads
 #' #...............................................................
 #'
-#' avail_cycles = available_RGTs(only_cycle_names = TRUE)
+#' avail_cycles = available_RGTs(only_cycle_names = TRUE,
+#'                               verbose = TRUE)
 #' avail_cycles
 #'
 #' choose_cycle = avail_cycles[3]
@@ -432,7 +626,7 @@ time_specific_orbits = function(date_from = NULL,
   if (verbose) t_start = proc.time()
   if (all(c(!is.null(date_from), !is.null(date_to), !is.null(RGT_cycle)))) stop("You have to specify either an RGT cycle ('RGT_cycle' parameter) or the 'date_from' and 'date_to' parameters!", call. = F)
 
-  avail_rgts = available_RGTs(only_cycle_names = FALSE)
+  avail_rgts = available_RGTs(only_cycle_names = FALSE, verbose = verbose)
   orbit_files = avail_rgts$orbit_files
   orbit_dates = avail_rgts$orbit_dates
   cycle_names = avail_rgts$names_cycles
@@ -640,5 +834,582 @@ time_specific_orbits = function(date_from = NULL,
   }
 
   return(all_cycles)
+}
+
+
+
+#' Utilizing Virtual File Systems (vsi) to extract the .kml from the .zip file
+#'
+#' @param icesat_rgt_url a character string specifying the input .zip URL
+#' @param download_zip a boolean. If TRUE the .zip file will be first downloaded and then the .kml files will be returned, otherwise the 'gdalinfo' function will be used as input to the R 'system2()' function to read the .kml files without downloading the .zip file. The 'gdalinfo' command requires that the user has configured GDAL properly.
+#' @param download_method a character string specifying the download method. Corresponds to the 'method' parameter of the 'utils::download.file()' function. Can be one of 'internal', 'wininet' (Windows only), 'libcurl', 'wget', 'curl' or 'auto'
+#' @param verbose a boolean. If TRUE then information will be printed out in the console
+#'
+#' @return an one column data.table of the output files
+#'
+#' @references
+#'
+#' https://icesat-2.gsfc.nasa.gov/science/specs
+#'
+#' https://gdal.org/user/virtual_file_systems.html
+#'
+#' @importFrom glue glue
+#' @importFrom utils download.file unzip
+#' @importFrom tools file_ext
+#' @importFrom sf gdal_utils
+#' @importFrom data.table setDT
+#'
+#' @export
+#'
+#' @examples
+#'
+#' \dontrun{
+#'
+#' require(IceSat2R)
+#'
+#' URL = 'https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/arcticallorbits.zip'
+#'
+#' #..................................
+#' # without downloading the .zip file
+#' #..................................
+#'
+#' res_out = vsi_kml_from_zip(icesat_rgt_url = URL,
+#'                            download_zip = FALSE,
+#'                            download_method = 'curl',
+#'                            verbose = TRUE)
+#' str(res_out)
+#'
+#'
+#' #.............................
+#' # by downloading the .zip file
+#' #.............................
+#'
+#' res_out = vsi_kml_from_zip(icesat_rgt_url = URL,
+#'                            download_zip = TRUE,
+#'                            download_method = 'curl',
+#'                            verbose = TRUE)
+#' str(res_out)
+#'
+#' }
+
+
+vsi_kml_from_zip = function(icesat_rgt_url,
+                            download_zip = FALSE,
+                            download_method = 'curl',
+                            verbose = FALSE) {
+
+  if (verbose) t_start = proc.time()
+  EXT_url = tools::file_ext(icesat_rgt_url)
+  url_pth = as.character(glue::glue('/vsizip/vsicurl/{icesat_rgt_url}'))
+
+  if (download_zip) {
+
+    tmp_url_file = tempfile(fileext = as.character(glue::glue('.{EXT_url}')))
+
+    if (verbose) cat(glue::glue("The '{icesat_rgt_url}' file will be downloaded (temporarily) in the '{tmp_url_file}' file ..."), '\n')
+    utils::download.file(url = icesat_rgt_url, destfile = tmp_url_file, method = download_method, quiet = !verbose)
+    info_url = utils::unzip(zipfile = tmp_url_file, list = TRUE, junkpaths = T)
+    info_url = info_url$Name
+    if (length(info_url) == 0) stop(glue::glue("The 'unzip -l' command didn't returned the required .kml url paths for the input url: {url_pth}!"), call. = F)
+  }
+  else {
+
+    info_url = sf::gdal_utils(util = 'info',
+                              source = url_pth,
+                              quiet = T)
+    if (info_url == "") {
+      if (verbose) message("The 'sf' gdalinfo returned an empty character string! Attempt to read the url using the OS configured 'gdalinfo' function ...")
+
+      info_url = suppressWarnings(tryCatch(system2(command = 'gdalinfo', args = url_pth, stdout = TRUE, stderr = FALSE), error = function(e) e))
+
+      if (inherits(info_url, 'error')) stop(glue::glue("The OS 'gdalinfo' gave the following error: {info_url$message}!"), call. = F)
+      if (length(info_url) == 0) stop(glue::glue("The OS 'gdalinfo' didn't returned the required .kml url paths for the gdal input url: {url_pth}!"), call. = F)
+    }
+  }
+
+  out_files = tools::file_ext(info_url)
+  tbl_files = data.frame(table(out_files))
+  tbl_files = tbl_files[order(tbl_files$Freq, decreasing = T), ]
+  type_files = as.character(tbl_files$out_files[1])
+  if (verbose) cat(glue::glue("The internal type of the .zip file is '{type_files}'"), '\n')
+
+  idx_kml = which(out_files == type_files)
+  if (length(idx_kml) == 0) stop("The .zip file (of the specified 'icesat_rgt_url' parameter) does not include any .kml files!", call. = F)
+  info_url = info_url[idx_kml]
+  info_url = trimws(info_url, which = 'both')
+
+  macosx_idx = which(gregexpr(pattern = 'MACOSX', text = info_url) != -1)
+  LEN_mc = length(macosx_idx)
+  if (LEN_mc > 0) {
+    if (verbose) message(glue::glue("{LEN_mc} files will be removed because they are hidden MACOSX files!"))
+    info_url = info_url[-macosx_idx]
+  }
+
+  LEN_url = length(info_url)
+  flag_error = F
+
+  if (type_files == 'kml') {
+    if (!LEN_url %in% c(1221, 1387)) {
+      flag_error = T
+      msg = "Except for the 'IS2_RGTs_cycle4_date_time_rev2_0.zip' file which has 1221, I expect that the remaining have 1387 ground track .kml files in every cycle (every cycle consists of approximately 91 days)!"
+    }
+  }
+
+  if (type_files == 'kmz') {
+    if (!LEN_url %in% c(7, 56)) {
+      flag_error = T
+      msg = "I expect that the 'antarcticaallorbits' and 'arcticallorbits' consist of 7 URLs ( 6 beams + 1 RGT ), whereas the 'ICESat2groundtracksWesternHem' and 'ICESat2groundtracksEasternHem' consist of 56 URLs ( 8 repeats * [6 beams + 1 RGT] )!"
+    }
+  }
+
+  if (flag_error) stop(msg, call. = F)
+
+  if (verbose) cat(glue::glue("The '{icesat_rgt_url}'  '{EXT_url}'  file includes  {LEN_url}  '{type_files}' files."), '\n')
+  if (download_zip) info_url = glue::glue("{url_pth}/{info_url}")
+  info_url = data.table::setDT(list(file = info_url))
+
+  if (verbose) compute_elapsed_time(time_start = t_start)
+
+  return(info_url)
+}
+
+
+
+#' Utilizing Virtual File Systems (vsi) and Well Known Text (WKT) to access nominal orbits
+#'
+#' @param orbit_area a character string specifying the earth partition to use, it can be one of 'antarctic', 'arctic', 'western_hemisphere' and 'eastern_hemisphere'
+#' @param track a character string specifying the orbit track. Can be one of 'GT1L','GT1R','GT2L','GT2R','GT3L','GT3R' or 'GT7'
+#' @param rgt_repeat an integer specifying the orbit repeat. This parameter defaults to 1 and it is relevant if a user chooses one of the 'western_hemisphere' or 'eastern_hemisphere' where there are 8 orbit repeats, whereas for the 'antarctic' and 'arctic' there is only 1 repeat
+#' @param wkt_filter either NULL, or a Well Known Text (WKT) character string to allow a user to restrict to an area of interest rather than processing all data
+#' @param download_zip a boolean. If TRUE the .zip file will be first downloaded and then the .kml files will be returned, otherwise the 'gdalinfo' function will be used as input to the R 'system2()' function to read the .kml files without downloading the .zip file. The 'gdalinfo' command requires that the user has configured GDAL properly. Set the parameter 'download_zip' to TRUE if GDAL is not (properly) installed.
+#' @param download_method a character string specifying the download method. Corresponds to the 'method' parameter of the 'utils::download.file()' function. Can be one of 'internal', 'wininet' (Windows only), 'libcurl', 'wget', 'curl' or 'auto'
+#' @param verbose a boolean. If TRUE then information will be printed out in the console
+#'
+#' @return an 'sf' object
+#'
+#' @references
+#'
+#' https://icesat-2.gsfc.nasa.gov/science/specs
+#'
+#' https://gdal.org/user/virtual_file_systems.html
+#'
+#' @importFrom glue glue
+#' @importFrom sf st_layers st_read st_as_sf
+#' @importFrom data.table rbindlist
+#'
+#' @export
+#'
+#' @examples
+#'
+#' \dontrun{
+#'
+#' require(IceSat2R)
+#' require(magrittr)
+#'
+#' #......................................
+#' # processing all data of the orbit area
+#' #......................................
+#'
+#' dat_rgt = vsi_nominal_orbits_wkt(orbit_area = 'eastern_hemisphere',
+#'                                  track = 'GT7',
+#'                                  rgt_repeat = 1,
+#'                                  wkt_filter = NULL,
+#'                                  download_method = 'curl',
+#'                                  download_zip = FALSE,
+#'                                  verbose = TRUE)
+#' str(dat_rgt)
+#'
+#'
+#' #...........................................
+#' # extracting nominal orbits only for the WKT
+#' #...........................................
+#'
+#' WKT = 'POLYGON ((-14.765 18.979, -11.25 18.979, -11.25 21.943, -14.765 21.943, -14.765 18.979))'
+#'
+#' dat_rgt = vsi_nominal_orbits_wkt(orbit_area = 'western_hemisphere',
+#'                                  track = 'GT3R',
+#'                                  rgt_repeat = 8,
+#'                                  wkt_filter = WKT,
+#'                                  download_method = 'curl',
+#'                                  download_zip = FALSE,
+#'                                  verbose = TRUE)
+#' str(dat_rgt)
+#' dat_rgt[[1]]$RGT                 # Reference Ground Tracks of input WKT
+#'
+#' #.............................
+#' # Visualize the results
+#' # (first compute the centroid)
+#' #.............................
+#'
+#' wkt_sf = sf::st_as_sfc(WKT, crs = 4326)
+#' centr_wkt = sf::st_coordinates(sf::st_centroid(wkt_sf))
+#'
+#' RGTs = mapview::mapview(dat_rgt, legend = F)
+#' AOI_wkt = mapview::mapview(wkt_sf, legend = F)
+#'
+#' lft = RGTs + AOI_wkt
+#' lft@map %>% leaflet::setView(lng = centr_wkt[, 'X'],
+#'                              lat = centr_wkt[, 'Y'],
+#'                              zoom = 7)
+#' }
+
+vsi_nominal_orbits_wkt = function(orbit_area,
+                                  track = 'GT7',
+                                  rgt_repeat = 1,
+                                  wkt_filter = NULL,
+                                  download_method = 'curl',
+                                  download_zip = FALSE,
+                                  verbose = FALSE) {
+
+  if (verbose) t_start = proc.time()
+  nomin_orb = available_nominal_orbits(orbit_area = orbit_area, verbose = verbose)
+  lst = list()
+
+  for (url_pth in nomin_orb) {
+
+    dat_out = vsi_kml_from_zip(icesat_rgt_url = url_pth,
+                               download_zip = download_zip,
+                               download_method = download_method,
+                               verbose = verbose)
+    lst[[url_pth]] = dat_out
+  }
+
+  if (verbose) cat('Data based on repeat and track will be kept ...\n')
+  dat_item = lapply(lst, function(x) {
+    idx = which(gregexpr(pattern = as.character(glue::glue('repeat{rgt_repeat}_{track}')), text = x$file) != -1)
+    if (length(idx) > 0) {
+      x[idx, , drop = F]
+    }
+    else {
+      NULL
+    }
+  })
+
+  all_null = all(as.vector(unlist(lapply(dat_item, function(x) is.null(x)))))
+  if (all_null) stop("The specified parameter setting (track, rgt_repeat) returned only NULL objects! Be aware: for the 'western_hemisphere' or 'eastern_hemisphere' there is max. 8 repeats, whereas for the 'antarctic' and 'arctic' is only 1 repeat!", call. = F)
+
+  dat_item = data.table::rbindlist(dat_item)
+  NROW = nrow(dat_item)
+
+  if (verbose) cat('Data based on repeat and track will be kept ...\n')
+  sublists = lapply(1:NROW, function(idx_row) {
+
+    if (verbose) cat(glue::glue("The file '{basename(dat_item$file[idx_row])}' will be processed ..."), '\n')
+    lrs = sf::st_layers(dsn = dat_item$file[idx_row])
+
+    NAMS = lrs$name
+    if (length(NAMS) == 0) stop(glue::glue("The '{dat_item$file[idx_row]}' does not include any layers!"), call. = F)
+    LEN_nams = length(NAMS)
+
+    dat_all = lapply(1:LEN_nams, function(x) {
+      LAYER = NAMS[x]
+      wkt_inp = character(0)
+      if (!is.null(wkt_filter)) wkt_inp = wkt_filter
+      lr_dat = sf::st_read(dsn = dat_item$file[idx_row], wkt_filter = wkt_inp, layer = LAYER, quiet = TRUE)
+      lr_dat
+    })
+
+    dat_all = data.table::rbindlist(dat_all)
+    dat_all = sf::st_as_sf(dat_all, crs = 4326)
+    dat_all$orbit_area = glue::glue("{rep(basename(dat_item$file[idx_row]), nrow(dat_all))}")
+    dat_all = dat_all[, c('Name', 'orbit_area')]
+    colnames(dat_all) = c('RGT', 'orbit_area', 'geometry')
+    dat_all
+  })
+
+  if (length(sublists) > 0) names(sublists) = basename(dat_item$file)
+
+  if (verbose) {
+    cat("Total ")
+    compute_elapsed_time(time_start = t_start)
+  }
+
+  return(sublists)
+}
+
+
+
+#' Utilizing Virtual File Systems (vsi) and Well Known Text (WKT) to access time specific orbits
+#'
+#' @param date_from a character string specifying the 'start' date in the format 'yyyy-MM-dd' (such as '2020-01-01')
+#' @param date_to a character string specifying the 'end' date in the format 'yyyy-MM-dd' (such as '2020-01-01')
+#' @param RGTs a character vector (consisting of one or more) Reference Ground Track (RGT). See the Examples section on how to come to these RGTs based on the "vsi_nominal_orbits_wkt()" function
+#' @param wkt_filter either NULL, or a Well Known Text (WKT) character string to allow a user to restrict to an area of interest rather than processing all data. It is possible that the WKT won't intersect with any of the available time specific orbits due to the sparsity of the coordinates (the output in that case will be an empty list)
+#' @param verbose a boolean. If TRUE then information will be printed out in the console
+#'
+#' @return a list of 'sf' objects where each sublist will represent a different RGT cycle
+#'
+#' @references
+#'
+#' https://icesat-2.gsfc.nasa.gov/science/specs
+#'
+#' https://gdal.org/user/virtual_file_systems.html
+#'
+#' @importFrom glue glue
+#' @importFrom sf st_layers st_read st_as_sf st_zm
+#' @importFrom data.table rbindlist data.table
+#' @importFrom utils flush.console
+#' @importFrom tools file_ext
+#' @importFrom lubridate yday
+#'
+#' @export
+#'
+#' @examples
+#'
+#' \dontrun{
+#'
+#' require(IceSat2R)
+#' require(magrittr)
+#'
+#' #...........................................
+#' # extracting nominal orbits only for the WKT
+#' #...........................................
+#'
+#' WKT = 'POLYGON ((-14.765 18.979, -11.25 18.979, -11.25 21.943, -14.765 21.943, -14.765 18.979))'
+#'
+#' dat_rgt = vsi_nominal_orbits_wkt(orbit_area = 'western_hemisphere',
+#'                                  track = 'GT3R',
+#'                                  rgt_repeat = 8,
+#'                                  wkt_filter = WKT,
+#'                                  download_method = 'curl',
+#'                                  download_zip = FALSE,
+#'                                  verbose = TRUE)
+#' str(dat_rgt)
+#'
+#' out_rgt = dat_rgt[[1]]$RGT
+#'
+#' #.........................................
+#' # time specific RGTs (for a time interval)
+#' # request using a single RGT cycle
+#' #.........................................
+#'
+#' date_start = '2020-01-01'
+#' date_end = '2020-02-01'
+#'
+#' orb_cyc_single = vsi_time_specific_orbits_wkt(date_from = date_start,
+#'                                               date_to = date_end,
+#'                                               RGTs = out_rgt,
+#'                                               wkt_filter = WKT,
+#'                                               verbose = TRUE)
+#' str(orb_cyc_single)
+#'
+#' #.........................................
+#' # time specific RGTs (for a time interval)
+#' # request using more than one RGT cycles
+#' #.........................................
+#'
+#' date_start = '2019-11-01'
+#' date_end = '2020-01-01'
+#'
+#' orb_cyc_multi = vsi_time_specific_orbits_wkt(date_from = date_start,
+#'                                              date_to = date_end,
+#'                                              RGTs = out_rgt,
+#'                                              wkt_filter = WKT,
+#'                                              verbose = TRUE)
+#' str(orb_cyc_multi)
+#' table(orb_cyc_multi$cycle)
+#'
+#'
+#' #.......................................................
+#' # visualization of the output cycles (including the WKT)
+#' #.......................................................
+#'
+#' orb_cyc_multi$cycle = as.factor(orb_cyc_multi$cycle)
+#' mp_orb = mapview::mapview(orb_cyc_multi, legend = TRUE, zcol = 'cycle')
+#'
+#' sf_aoi = sf::st_as_sfc(WKT, crs = 4326)
+#' mp_aoi = mapview::mapview(sf_aoi, alpha.regions = 0.3, legend = F)
+#'
+#' mp_orb + mp_aoi
+#'
+#' }
+
+vsi_time_specific_orbits_wkt = function(date_from,
+                                        date_to,
+                                        RGTs,
+                                        wkt_filter = NULL,
+                                        verbose = FALSE) {
+  if (verbose) t_start = proc.time()
+
+  RGTs = as.vector(sapply(RGTs, function(x) {                     # make sure to add the required digits
+    if (nchar(x) == 1) x = glue::glue("000{x}")
+    else if (nchar(x) == 2) x = glue::glue("00{x}")
+    else if (nchar(x) == 3) x = glue::glue("0{x}")
+    else x
+
+    as.character(x)
+  }))
+
+  avail_rgts = available_RGTs(only_cycle_names = FALSE, verbose = verbose)
+  orbit_files = avail_rgts$orbit_files
+  orbit_dates = avail_rgts$orbit_dates
+
+  seq_dates = seq(from = as.Date(date_from), to = as.Date(date_to), by = '1 day')
+
+  inters_dates = as.vector(unlist(lapply(orbit_dates, function(x) length(intersect(x = x, y = seq_dates)))))
+
+  res_lst = list()
+  idx_inters = which(inters_dates > 0)
+  LEN_inters = length(idx_inters)
+
+  if (LEN_inters > 0) {
+
+    cycles_subs = names(orbit_files)[idx_inters]
+    zip_subs = as.vector(unlist(orbit_files))[idx_inters]
+
+    if (verbose) {
+      cat(glue::glue("In total there are {LEN_inters} intersected dates for which data will be processed!"), '\n')
+      cat(glue::glue("The RGT cycles from which data will be processed are: {paste(cycles_subs, collapse = ', ')}"), '\n')
+    }
+
+    for (item_cy in seq_along(zip_subs)) {
+
+      iter_cycle = cycles_subs[item_cy]
+      if (verbose) {
+        cat("-------------------------------------------------\n")
+        cat(glue::glue("RGTs of cycle '{iter_cycle}' will be processed ..."), '\n')
+        cat("-------------------------------------------------\n")
+      }
+
+      iter_url = zip_subs[item_cy]
+      zip_dat = vsi_kml_from_zip(icesat_rgt_url = iter_url,
+                                 download_zip = FALSE,
+                                 download_method = 'curl',
+                                 verbose = TRUE)
+      if (nrow(zip_dat) > 0) {
+
+        greg_expr = as.character(glue::glue("IS2_RGT_{RGTs}"))
+        rows_inters = as.vector(sapply(greg_expr, function(x) which(gregexpr(pattern = x, text = zip_dat$file) != -1)))
+        if (length(rows_inters)) {
+          zip_dat_subs = zip_dat[rows_inters, , drop = F]
+          NROW = nrow(zip_dat_subs)
+
+          sublists = lapply(1:NROW, function(idx_row) {
+
+            if (verbose) {
+              message(glue::glue("{iter_cycle} FILE: "), idx_row, "/", NROW, "\r", appendLF = FALSE)
+              utils::flush.console()
+            }
+
+            lrs = sf::st_layers(dsn = zip_dat_subs$file[idx_row])
+
+            NAMS = lrs$name
+            if (length(NAMS) == 0) stop(glue::glue("The '{zip_dat_subs$file[idx_row]}' does not include any layers!"), call. = F)
+            LEN_nams = length(NAMS)
+
+            dat_all = lapply(1:LEN_nams, function(x) {
+              LAYER = NAMS[x]
+              wkt_inp = character(0)
+              if (!is.null(wkt_filter)) wkt_inp = wkt_filter
+              lr_dat = sf::st_read(dsn = zip_dat_subs$file[idx_row], wkt_filter = wkt_inp, layer = LAYER, quiet = TRUE)
+              lr_dat
+            })
+
+            n_rows = as.vector(unlist(lapply(dat_all, nrow)))
+            idx_relev = which(n_rows > 0)
+
+            if (length(idx_relev) > 0) {
+              dat_all = dat_all[idx_relev]
+
+              dat_all = data.table::rbindlist(dat_all)
+
+              if (nrow(dat_all) > 0) {
+                dat_all
+              }
+              else {
+                NULL
+              }
+            }
+          })
+
+          nams_init = basename(zip_dat_subs$file)                           # 'nams_init' will be modified if there are NULL sublists
+          idx_null = which(as.vector(unlist(lapply(sublists, is.null))))
+          if (length(idx_null) > 0) {
+            if (verbose) message(glue::glue("{length(idx_null)} out of {length(sublists)} sublists were empty and will be removed!"))
+            sublists = sublists[-idx_null]
+            nams_init = nams_init[-idx_null]
+          }
+
+          if (length(sublists) > 0) {
+            names(sublists) = nams_init
+            res_lst[[iter_cycle]] = sublists
+          }
+        }
+      }
+    }
+  }
+
+  LEN_lst = length(res_lst)
+
+  if (LEN_lst > 0) {
+
+    nams_out = names(res_lst)
+    if (verbose) cat(glue::glue("In total {LEN_lst} RGT cycles will be included in the output 'sf' object ({paste(nams_out, collapse = ', ')})!"), '\n')
+
+    res_lst = lapply(1:LEN_lst, function(y) {
+
+      x = res_lst[[y]]
+
+      nams_lst = names(x)
+      unq_ext = unique(tools::file_ext(nams_lst))
+      if (length(unq_ext) > 1) stop("I expect either a '.kml' or '.kmz' file extension!", call. = F)
+
+      if (verbose) cat(glue::glue("output of '{nams_out[y]}' will be re-formatted ..."), '\n')
+
+      x = data.table::rbindlist(x)
+      x = sf::st_as_sf(x, crs = 4326)              # by default use 'EPSG:4326'
+
+      descr_x = x$description
+      descr_x_invalid = (any(descr_x == "") | any(is.na(descr_x)))
+
+      if (descr_x_invalid) {
+
+        dat_tim = gsub(glue::glue(".{unq_ext}"), '', nams_lst)       # remove the file extension from the file names
+        dat_tim = lapply(strsplit(dat_tim, '_'), function(k) {
+          iter_date = k[length(k)]
+          iter_cycle = k[length(k) - 1]
+          iter_cycle = as.integer(gsub('[a-zA-Z]', '', iter_cycle))
+          list(date = iter_date, cycle = iter_cycle)
+        })
+        dat_tim = data.table::rbindlist(dat_tim)
+        yday = strptime(x = dat_tim$date, format = "%d-%b-%Y")
+        dat_tim$date = as.character(glue::glue("{dat_tim$date} 00:00:00"))       # add time so that it matches the next "else" condition time-format
+
+        x$description = NULL
+        x$RGT = as.integer(gsub('RGT ', '', x$Name))
+        x$Date_time = dat_tim$date                                               # add the date
+        x$day_of_year = lubridate::yday(x = yday)
+        x$cycle = dat_tim$cycle
+      }
+      else {
+        if (verbose) cat("The 'description' column of the output data will be processed ...\n")
+        descr_proc = strsplit(x = x$description, split = '\n')
+        descr_proc = data.table::data.table(do.call(rbind, descr_proc), stringsAsFactors = F)
+        colnames(descr_proc) = c('RGT', 'Date_time', 'DOY', 'cycle')
+
+        x$description = NULL
+        x$RGT = as.integer(gsub('RGT ', '', descr_proc$RGT))                           # keep the integer from the character string which corresponds to the 'RGT'
+        x$Date_time = descr_proc$Date_time
+        x$day_of_year = as.integer(gsub('DOY ', '', descr_proc$DOY))                   # keep the integer from the character string which corresponds to the 'DOY'
+        x$cycle = as.integer(gsub('Cycle ', '', descr_proc$cycle))                     # keep the integer from the character string which corresponds to the 'cycle'
+      }
+
+      x = sf::st_zm(x = x, drop = T)                                                   # drop the Z dimension
+      x
+    })
+
+    res_lst = sf::st_as_sf(data.table::rbindlist(res_lst))
+    res_lst$Date_time = strptime(x = res_lst$Date_time, format = "%d-%b-%Y %H:%M:%S")          # convert to Date-Time ('POSIXlt') format  [ !! I have to do that outside the for-loop, otherwise I might receive an error ]
+  }
+  else {
+    if (verbose) message(glue::glue("The specified parameter setting returned an empty list for all {length(idx_inters)} processed cycles!"))
+  }
+
+  if (verbose) {
+    cat("Total ")
+    compute_elapsed_time(time_start = t_start)
+  }
+
+  return(res_lst)
 }